weather bot/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
# For Kalshi authenticated orders (RSA-PSS signing):
cryptography>=41.0.0
